Upgrading data types

+++ Source/IP_Chol_tdfs.c
@@ -12,18 +12,18 @@
 
 /* Purpose: Depth-first search and postorder of a tree rooted at node j */
 
-int IP_Chol_tdfs 
+int64_t IP_Chol_tdfs 
 (
-    int j,      // Root node
-    int k,      
-    int* head,  // Head of list
-    int* next,  // Next node in the list
-    int* post,  // Post ordered tree
-    int* stack  // Stack of nodes
+    int64_t j,      // Root node
+    int64_t k,      
+    int64_t* head,  // Head of list
+    int64_t* next,  // Next node int64_t* the list
+    int64_t* post,  // Post ordered tree
+    int64_t* stack  // Stack of nodes
 )
 { 
-    int i, p, top = 0 ;
-    if (!head || !next || !post || !stack) return (-1) ;    /* check inputs */
+    int64_t i, p, top = 0 ;
+    if (!head || !next || !post || !stack) return (-1) ;    /* check int64_t*puts */
     stack [0] = j ;                 /* place j on the stack */
     while (top >= 0)                /* while (stack is not empty) */
     {
